+++ omnistrate_tests/test_upgrade_version.py
@@ -55,7 +55,7 @@ parser.add_argument("--rdb-config", required=False, default="medium")
 parser.add_argument("--aof-config", required=False, default="always")
 parser.add_argument("--host-count", required=False, default="6")
 parser.add_argument("--cluster-replicas", required=False, default="1")
-
+parser.add_argument("--debug",required=False,default=False)
 
 parser.set_defaults(tls=False)
 args = parser.parse_args()
@@ -70,8 +70,9 @@ def signal_handler(sig, frame):
     sys.exit(0)
 
 
-signal.signal(signal.SIGINT, signal_handler)
-signal.signal(signal.SIGTERM, signal_handler)
+if args.debug is False:
+    signal.signal(signal.SIGTERM, signal_handler)
+    signal.signal(signal.SIGINT, signal_handler)
 
 
 def test_upgrade_version():
@@ -185,7 +186,8 @@ def test_upgrade_version():
         query_data(instance)
     except Exception as e:
         logging.exception(e)
-        instance.delete(False)
+        if args.debug is False:
+            instance.delete(False)
         raise e
 
     # 7. Delete the instance
